# src/main/resources/public/docs/v1/openapi.yml
openapi: "3.0.1"
info:
  title: Attoly
  version: v1
  x-logo:
    url: /site/img/logo.svg
    altText: Attoly Logo
  description: |
    This is the backend of the Attoly URL shortener platform, responsible for
    data storage and exchange. Attoly allows creating short links as aliases
    for long URLs, which are easier to embed in websites, chats and documents.
    
    # Introduction
    This is the official documentation of the Attoly RESTful API. Based on
    simple REST principles, the Attoly RESTful API endpoints return JSON
    metadata about created and managed short link URL mappings aka shortcuts.
    This API also provides access to user related data, like account information
    and your personal shortcut library.
    
    # Overview
    The following settings and techniques apply to the entire Attoly RESTful API.
    
    ## HTTP Verbs
    There are a couple of possible HTTP verbs used for the RESTful API. The following
    table clearly clarifies their use and meaning within the API.
    
    | Verb         | Description                                                       |
    |--------------|-------------------------------------------------------------------|
    | HEAD         | Can be ran against any resource to get just the HTTP header info. |
    | GET          | Used for retrieving resources.                                    |
    | POST         | Used for creating resources.                                      |
...
    
    ## Timestamps
...
    
...
    
...

...
servers:
...
paths:
...
  /shortcuts/{tag}/reports:
    post:
      tags:
        - Report
      summary: Reports a shortcut.
      description: >
        Reports a shortcut based on its tag. The report is sent to the
        administrator of the service. The administrator can then decide
        whether to delete the shortcut or not.
      parameters:
        - name: tag
          in: path
          description: Unique tag of the shortcut.
          required: true
          schema:
            type: string
            example: hluEinvI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreationDto"
            example:
              reason: SPAM
              comment: Contains offensive content.
      responses:
        204:
          description: Successfully reported the shortcut.
        404:
          $ref: "#/components/responses/NotFound"
        401:
          $ref: "#/components/responses/Unauthenticated"
components:
  schemas:
    ErrorDto:
      type: object
      properties:
        message:
          type: string
          description: The human-readable error message.
        status:
          type: integer
          description: The HTTP status code number.
        timestamp:
          type: string
          format: date-time
          description: The time the error occurred.
        path:
          type: string
          description: The endpoint path on which the error occurred.
        details:
          description: Further information on the error or its cause.
          type: array
          items:
            type: object
      required:
        - message
        - status
        - timestamp
        - path
...
    ReportCreationDto:
      type: object
      properties:
        reason:
          type: string
          enum: [ SPAM, MALWARE, PHISHING, DEFACEMENT ]
          description: The reason for reporting the shortcut
        comment:
          type: string
          description: A comment explaining the reason in more detail.
          maxLength: 2000
          minLength: 1
      required:
        - reason
        - comment
  responses:
...
    NotFound:
      description: The specified resource was not found.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDto"
...
    Unauthenticated:
      description: The access must be authenticated.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorDto"
...
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: jwt
security:
  - jwt: [ ]